refactor(tsz-core): folderize embedded_libs module
